Complete Defense Charlies and change hmac_token to transfer_token

+++ proj2/code/router.js
@@ -30,15 +30,28 @@ function sanitized_object(object) {
 }
 
 // Defense Bravo
-// Change server secret key every 5 minutes
-const secret_key_timeout_ms = 300000
-var server_secret_key = generateRandomness()
-setInterval(function() {server_secret_key = generateRandomness()}, secret_key_timeout_ms);
-function generate_hmac_token(account) {
+// Change transfer server secret key every 5 minutes
+const transfer_secret_key_timeout_ms = 300000;
+var transfer_server_secret_key = generateRandomness();
+setInterval(function() {transfer_server_secret_key = generateRandomness()}, transfer_secret_key_timeout_ms);
+function generate_transfer_token(account) {
   if (typeof account === 'undefined' || account === null) return account;
-  return HMAC(server_secret_key, account.username.concat(account.hashedPassword));
+  return HMAC(transfer_server_secret_key, account.username.concat(account.hashedPassword));
 }
 
+// Defense Charlie
+var login_session_signature = null;
+const session_server_secret_key = generateRandomness();
+function generate_session_signature(session) {
+  if (typeof session.account.username === 'undefined' || session.account.username === null) return null;
+  return HMAC(session_server_secret_key, session.account.username.concat(session.account.hashedPassword));
+}
+function force_logout(req, res, next) {
+  req.session.loggedIn = false;
+  req.session.account = {};
+  login_session_signature = null;
+  render(req, res, next, 'index', 'Bitbar Home', 'Logged out due to a security reason! Please login again.');
+}
 
 function render(req, res, next, page, title, errorMsg = false, result = null) {
   res.render(
@@ -55,11 +68,18 @@ function render(req, res, next, page, title, errorMsg = false, result = null) {
 
 
 router.get('/', (req, res, next) => {
+  if(generate_session_signature(req.session) !== login_session_signature) {
+    force_logout(req, res, next);
+  }
   render(req, res, next, 'index', 'Bitbar Home');
 });
 
 
 router.post('/set_profile', asyncMiddleware(async (req, res, next) => {
+  if(generate_session_signature(req.session) !== login_session_signature) {
+    force_logout(req, res, next);
+  }
+
   req.session.account.profile = req.body.new_profile;
   console.log(req.body.new_profile);
   const db = await dbPromise;
@@ -84,6 +104,7 @@ router.get('/get_login', asyncMiddleware(async (req, res, next) => {
       await sleep(2000);
       req.session.loggedIn = true;
       req.session.account = result;
+      login_session_signature = generate_session_signature(req.session);
       render(req, res, next, 'login/success', 'Bitbar Home');
       return;
     }
@@ -121,6 +142,7 @@ router.post('/post_register', asyncMiddleware(async (req, res, next) => {
     profile: '',
     bitbars: 100,
   };
+  login_session_signature = generate_session_signature(req.session);
   render(req, res, next,'register/success', 'Bitbar Home');
 }));
 
@@ -130,11 +152,17 @@ router.get('/close', asyncMiddleware(async (req, res, next) => {
     render(req, res, next, 'login/form', 'Login', 'You must be logged in to use this feature!');
     return;
   };
+
+  if(generate_session_signature(req.session) !== login_session_signature) {
+    force_logout(req, res, next);
+  }
+
   const db = await dbPromise;
   const query = `DELETE FROM Users WHERE username == "${req.session.account.username}";`;
   await db.get(query);
   req.session.loggedIn = false;
   req.session.account = {};
+  login_session_signature = null;
   render(req, res, next, 'index', 'Bitbar Home', 'Deleted account successfully!');
 }));
 
@@ -142,6 +170,7 @@ router.get('/close', asyncMiddleware(async (req, res, next) => {
 router.get('/logout', (req, res, next) => {
   req.session.loggedIn = false;
   req.session.account = {};
+  login_session_signature = null;
   render(req, res, next, 'index', 'Bitbar Home', 'Logged out successfully!');
 });
 
@@ -151,6 +180,10 @@ router.get('/profile', asyncMiddleware(async (req, res, next) => {
     render(req, res, next, 'login/form', 'Login', 'You must be logged in to use this feature!');
     return;
   };
+
+  if(generate_session_signature(req.session) !== login_session_signature) {
+    force_logout(req, res, next);
+  }
 
   if(req.query.username != null) { // if visitor makes a search query
     const db = await dbPromise;
@@ -178,8 +211,13 @@ router.get('/transfer', (req, res, next) => {
     render(req, res, next, 'login/form', 'Login', 'You must be logged in to use this feature!');
     return;
   };
+
+  if(generate_session_signature(req.session) !== login_session_signature) {
+    force_logout(req, res, next);
+  }
+
   render(req, res, next, 'transfer/form', 'Transfer Bitbars', false,
-         {receiver:null, amount:null, hmac_token:generate_hmac_token(req.session.account)});
+         {receiver:null, amount:null, transfer_token:generate_transfer_token(req.session.account)});
 });
 
 
@@ -189,16 +227,20 @@ router.post('/post_transfer', asyncMiddleware(async(req, res, next) => {
     return;
   };
 
+  if(generate_session_signature(req.session) !== login_session_signature) {
+    force_logout(req, res, next);
+  }
+
   if(req.body.destination_username === req.session.account.username) {
     render(req, res, next, 'transfer/form', 'Transfer Bitbars', 'You cannot send money to yourself!',
-           {receiver:null, amount:null, hmac_token:generate_hmac_token(req.session.account)});
+           {receiver:null, amount:null, transfer_token:generate_transfer_token(req.session.account)});
     return;
   }
 
-  if(req.body.hmac_token !== generate_hmac_token(req.session.account)) {
+  if(req.body.transfer_token !== generate_transfer_token(req.session.account)) {
     render(req, res, next, 'transfer/form', 'Transfer Bitbars',
-           'Transfer Session Expired! Please try again within '.concat(secret_key_timeout_ms/60000).concat(" minutes"),
-           {receiver:null, amount:null, hmac_token:generate_hmac_token(req.session.account)});
+           'Transfer Session Expired! Please try again within '.concat(transfer_secret_key_timeout_ms/60000).concat(" minutes"),
+           {receiver:null, amount:null, transfer_token:generate_transfer_token(req.session.account)});
     return;
   }
 
@@ -209,7 +251,7 @@ router.post('/post_transfer', asyncMiddleware(async(req, res, next) => {
     const amount = parseInt(req.body.quantity);
     if(Number.isNaN(amount) || amount > req.session.account.bitbars || amount < 1) {
       render(req, res, next, 'transfer/form', 'Transfer Bitbars', 'Invalid transfer amount!',
-             {receiver:null, amount:null, hmac_token:generate_hmac_token(req.session.account)});
+             {receiver:null, amount:null, transfer_token:generate_transfer_token(req.session.account)});
       return;
     }
 
@@ -220,7 +262,7 @@ router.post('/post_transfer', asyncMiddleware(async(req, res, next) => {
     query = `UPDATE Users SET bitbars = "${receiverNewBal}" WHERE username == "${receiver.username}";`;
     await db.exec(query);
     render(req, res, next, 'transfer/success', 'Transfer Complete', false,
-           {receiver, amount, hmac_token:generate_hmac_token(req.session.account)});
+           {receiver, amount, transfer_token:generate_transfer_token(req.session.account)});
   } else { // user does not exist
     let q = req.body.destination_username;
     if (q == null) q = '';
@@ -231,7 +273,7 @@ router.post('/post_transfer', asyncMiddleware(async(req, res, next) => {
       q = q.replace(/script|SCRIPT|img|IMG/g, '');
     }
     render(req, res, next, 'transfer/form', 'Transfer Bitbars', `User ${q} does not exist!`,
-           {receiver:null, amount:null, hmac_token:generate_hmac_token(req.session.account)});
+           {receiver:null, amount:null, transfer_token:generate_transfer_token(req.session.account)});
   }
 }));
